8.USER PROFILE-11 Video Owner ~ 9.WEBPACK - 5 MiniCssExtractPlugin

+++ src/controllers/videoController.js
@@ -1,8 +1,9 @@
 import Video from "../models/Video";
+import User from "../models/User";
 
 export const home = async (req, res) => {
   try {
-    const videos = await Video.find({});
+    const videos = await Video.find({}).sort({ createdAt: "desc" });
     return res.render("home", {
       pageTitle: "Home",
       videos,
@@ -14,7 +15,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     res.status(404).render("404", { pageTitle: "Video not found" });
   }
@@ -27,9 +28,10 @@ export const watch = async (req, res) => {
 export const getEdit = async (req, res) => {
   const { id } = req.params;
   const video = await Video.findById(id);
-  if (!video) {
-    res.status(404).render("404", { pageTitle: "Video not found" });
-  }
+  if (!video)
+    return res.status(404).render("404", { pageTitle: "Video not found" });
+  if (String(video.owner) !== String(req.session.user._id))
+    return res.status(403).redirect("/");
   return res.render("edit", { pageTitle: `Edit : ${video.title}`, video });
 };
 
@@ -37,9 +39,9 @@ export const postEdit = async (req, res) => {
   const { id } = req.params;
   const { title, description, hashtags } = req.body;
   const video = await Video.exists({ _id: id });
-  if (!video) {
-    res.render("404", { pageTitle: "Video not found" });
-  }
+  if (!video) return res.render("404", { pageTitle: "Video not found" });
+  if (String(video.owner) !== String(req.session.user._id))
+    return res.status(403).redirect("/");
   await Video.findByIdAndUpdate(id, {
     title,
     description,
@@ -54,15 +56,20 @@ export const getUpload = (req, res) => {
 export const postUpload = async (req, res) => {
   try {
     const {
+      session: { user: _id },
       body: { title, description, hashtags },
       file: { path: fileUrl },
     } = req;
-    await Video.create({
+    const newVideo = await Video.create({
       fileUrl,
       title,
       description,
       hashtags: Video.hashtagsForm(hashtags),
+      owner: _id,
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (error) {
     return res.status(400).render("upload", { errmsg: error._message });
@@ -71,6 +78,15 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
   const { id } = req.params;
+  const video = await Video.findById(id);
+  if (!video) return res.render("404", { pageTitle: "Video not found" });
+  if (String(video.owner) !== String(req.session.user._id))
+    return res.status(403).redirect("/");
+  const user = await User.findById(video.owner);
+  user.videos = user.videos.filter(
+    (userVideo) => String(userVideo) !== String(video._id),
+  );
+  user.save();
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
